requirements.txt
asgiref==3.8.1
coverage==7.8.2
dj-database-url==2.3.0
Django==5.2
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.10
exceptiongroup==1.2.2
factory_boy==3.3.3
Faker==37.1.0
gunicorn==23.0.0
inflection==0.5.1
iniconfig==2.1.0
packaging==25.0
pillow==11.1.0
pluggy==1.5.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
pytest==8.3.5
pytest-django==4.11.1
python-dotenv==1.1.0
pytz==2025.1
PyYAML==6.0.2
setuptools==80.8.0
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
whitenoise==6.9.0
